ntp_client.pm: only apply workaround on SLE
follow-up fixes after #17562

+++ tests/console/ntp_client.pm
@@ -47,7 +47,7 @@ sub run {
     systemctl 'status chronyd';
 
     # due to bsc#1214141, we need to remove and install again chrony and chrony-pool-suse (just reinstalling doesn't work)
-    if (script_run('cat /etc/chrony.d/pool.conf | grep -q ^pool') != 0) {
+    if (is_sle() && script_run('cat /etc/chrony.d/pool.conf | grep -q ^pool') != 0) {
         record_info 'workaround for bsc#1214141';
         zypper_call 'rm chrony-pool-suse';
         zypper_call 'in chrony-pool-suse';
